v1.0

+++ src/install.sh
@@ -80,7 +80,7 @@ DNSOverTLS=yes
 EOF
 fi
 
-sudo ln -sf /run/systemd/resolve/stub-resolv.conf /etc/resolv.conf
+[ -e /run/systemd/resolve/stub-resolv.conf ] && sudo ln -sf /run/systemd/resolve/stub-resolv.conf /etc/resolv.conf
 
 sudo systemctl restart systemd-resolved
